pyproject.toml
[project]
name = "trash-classification-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.4"
dependencies = [
    "chromadb>=0.6.0",
    "ipykernel>=7.2.0",
    "markupsafe==3.0.3",
    "matplotlib>=3.10.8",
    "open-clip-torch>=2.24.0",
    "pandas",
    "pillow",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "sentence-transformers>=3.0.0",
    "torch>=2.10.0",
    "tqdm",
    "torchvision>=0.25.0",
    "transformers",
]

[tool.uv]
package = true

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"

[tool.uv.sources]
torch = { index = "pytorch-cu130" }
torchvision = { index = "pytorch-cu130" }
